requirements.txt
asgiref==3.8.1
channels==4.2.0
Django==5.1.4
django_channels==0.7.0
djangorestframework_simplejwt==5.3.1
PyJWT==2.10.1
PyJWT==2.10.1
Requests==2.32.3
